requirements.txt
streamlit
python-docx
sentence-transformers
faiss-cpu
transformers
torch
numpy